+++ src/main/java/me/myogoo/ae2tb/mixin/jei/CombinedRecipeFocusSourceMixin.java
@@ -1,5 +1,6 @@
 package me.myogoo.ae2tb.mixin.jei;
 
+import me.myogoo.ae2tb.client.KeyBindings;
 import mezz.jei.common.input.IInternalKeyMappings;
 import mezz.jei.gui.input.CombinedRecipeFocusSource;
 import mezz.jei.gui.input.UserInput;
@@ -17,6 +18,9 @@ public class CombinedRecipeFocusSourceMixin {
             remap = false
     )
     private static void ae2tb$isConflictingVanillaMouseButton(UserInput input, IInternalKeyMappings keyBindings, CallbackInfoReturnable<Boolean> cir) {
-        cir.setReturnValue(cir.getReturnValue() || input.getKey().getName().equals("key.mouse.middle"));
+        cir.setReturnValue(cir.getReturnValue()
+                || input.is(KeyBindings.PICKUP_SINGLE_ITEM)
+                || input.is(KeyBindings.PICKUP_SET_ITEM)
+                || input.is(KeyBindings.PICKED_ITEM_AUTOCRAFTING));
     }
 }