+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Features
 - Added common QoL config.
-    - `allowBookmarkInteractionWithoutUpgrade`: removes upgrade card requirements.
+    - `QoL`: allows bookmark interaction and bookmark amount rendering without the Terminal Bookmark Interact Card.
 
 ## Versions
 - Updated Myotus to `19.1.0`.